clean up & refactoring
