--- packages/components/icons/activity-type-icons/index.ts
@@ -1,4 +1,0 @@
-export * from '@acter/components/icons/activity-type-icons/event-activity-icon'
-export * from '@acter/components/icons/activity-type-icons/idea-activity-icon'
-export * from '@acter/components/icons/activity-type-icons/meeting-activity-icon'
-export * from '@acter/components/icons/activity-type-icons/project-activity-icon'
